requirements.txt
requests
flask
pyngrok
python-dotenv
